requirements.txt
attrs==20.3.0
aws-cdk.assets==1.94.1
aws-cdk.aws-apigateway==1.94.1
aws-cdk.aws-applicationautoscaling==1.94.1
aws-cdk.aws-autoscaling-common==1.94.1
aws-cdk.aws-certificatemanager==1.94.1
aws-cdk.aws-cloudformation==1.94.1
aws-cdk.aws-cloudwatch==1.94.1
aws-cdk.aws-codeguruprofiler==1.94.1
aws-cdk.aws-cognito==1.94.1
aws-cdk.aws-dynamodb==1.94.1
aws-cdk.aws-ec2==1.94.1
aws-cdk.aws-ecr==1.94.1
aws-cdk.aws-ecr-assets==1.94.1
aws-cdk.aws-efs==1.94.1
aws-cdk.aws-elasticloadbalancingv2==1.94.1
aws-cdk.aws-events==1.94.1
aws-cdk.aws-iam==1.94.1
aws-cdk.aws-kinesis==1.94.1
aws-cdk.aws-kms==1.94.1
aws-cdk.aws-lambda==1.94.1
aws-cdk.aws-lambda-event-sources==1.94.1
aws-cdk.aws-logs==1.94.1
aws-cdk.aws-route53==1.94.1
aws-cdk.aws-s3==1.94.1
aws-cdk.aws-s3-assets==1.94.1
aws-cdk.aws-s3-notifications==1.94.1
aws-cdk.aws-signer==1.94.1
aws-cdk.aws-sns==1.94.1
aws-cdk.aws-sns-subscriptions==1.94.1
aws-cdk.aws-sqs==1.94.1
aws-cdk.aws-ssm==1.94.1
aws-cdk.cloud-assembly-schema==1.94.1
aws-cdk.core==1.94.1
aws-cdk.custom-resources==1.94.1
aws-cdk.cx-api==1.94.1
aws-cdk.region-info==1.94.1
cattrs==1.3.0
constructs==3.3.68
jsii==1.25.0
path==15.1.2
Pillow==8.1.2
publication==0.0.3
python-dateutil==2.8.1
six==1.15.0
typing-extensions==3.7.4.3
